Major speed improvement and bug fixes
Major change: The widget no longer makes the API request every time it
reloads, now you must tap the timer button to reload the Todoist tasks
and Toggl timers. This significantly reduces loading times.

Added a blank task to allow a timer with no description. Set this timer
as the default timer.

Fixed a bug that caused users to be unable to select the first item on
pages beyond page one.

Fixed the very lazy hardcoded number of lines in the task selection
view. Now dynamic with widget size.